make bigint

+++ python_modules/dagster-graphql/dagster_graphql/schema/pipelines/pipeline.py
@@ -248,7 +248,7 @@ class GrapheneAsset(graphene.ObjectType):
         cursor=graphene.String(),
     )
     definition = graphene.Field("dagster_graphql.schema.asset_graph.GrapheneAssetNode")
-    latestEventSortKey = graphene.Field(graphene.Int)
+    latestEventSortKey = graphene.Field(graphene.BigInt)
 
     class Meta:
         name = "Asset"
